Changes to fix Bender bug (#102)
This set of changes removes a multi upload part that fails for whatever reason during an UploadPart operation from the map of uploads. It also causes the handler to fail faster by pausing the streaming as soon as an event is found to fail.

We've validated that this change results in not seeing the "AmazonS3Exception: The XML you provided was not well-formed" stacktraces anymore and that the issue is the ResetException detailed above.